refactor(eslint): split local rules by role
